Reorganize e2e test directory
