+++ e2e/tenant/database/sqlc.yaml
